Add some submodules for optional (possibly nonpublic) features
We branch in most of the VPC project files.  These are not useful to the opensource
code without the partner module.  They are only needed to build SDR or console support.
But if I don't branch the files then I have to leave a symlink for each one, and since
they are small and don't have anything secret, it's easier to just have the actual files
themselves.

We branch in some third party sources where we need to control the
version and/or need to make modifications.

We add symlinks to directories in the modules folder.  These will be broken if you
don't have access to that stuff, but it's the cleanest way I know to "overlay" one
project on top of another.

Also add some public header files which are not part of the Steamworks SDK right now
but might be in the future.

I really wish there was a way for me to NOT pin the submodule to a particular
commit, and just always have it point at whatever the current head is.

I used a .gitattributes file as a workaround for Windows directory symlinks

https://github.com/git-for-windows/git/issues/2177